requirements.txt
Pillow==8.1.0
urllib3==1.26.3